Fixed test failure in BioTokenTest.  Missing cast from char to int so it would be "boxed" correctly.
git-svn-id: svn+ssh://svn.code.sf.net/p/pcgen/code/Trunk@1155 ef53c05b-3124-4d59-ae7d-fe3273f42be7

+++ code/src/java/pcgen/io/FileAccess.java
@@ -149,7 +149,7 @@ public final class FileAccess
 			for (int i = 0; i < aString.length(); i++)
 			{
 				final char c = aString.charAt(i);
-				final String xlation = outputFilter.get(c);
+				final String xlation = outputFilter.get((int)c);
 
 				if (xlation != null)
 				{
